P4ADEV-2177 updated swagger and added data-exports controller
